pyproject.toml
[project]
name = "gitlab-data-docs-cn"
version = "0.1.0"
description = "GitLab 企业数据平台文档中文翻译"
authors = [
    { name = "wyaming", email = "wyaming89@gmail.com" }
]
dependencies = [
    "python-dotenv==1.0.0",
    "httpx==0.27.0",
    "markdown==3.5.2",
    "tqdm==4.66.1",
    "beautifulsoup4==4.12.2",
    "html2text==2020.1.16",
    "aiofiles==23.2.1",
    "openai>=1.68.2",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py38" 
